Implement a more complete grouping outline
Currently, the only implemented outlining algorithm is using the
first and last pixel that were discovered in the grouping. This
only results in a very rough outline though.

This commit adds a new implementation for a more complete outline
that uses the most extreme pixel on each side of the grouping. We
know that the first pixel will be the topmost pixel and the last
pixel will be the bottommost pixel. For the leftmost and rightmost
we iterate through all pixels and store the pixels at the two
extremes.