Allow image to be resized before it is analyzed
Introduce the image resize options in ImageAnalyzer to handle the
analysis of large images. This will reduce the accuracy of the
analysis.

The resize dimensions that can be provided are bounds, an image
will keep its aspect ratio when resized.

No external libraries are used.

The UI additions for this feature will be added at a later date.